+++ src/fx/never_complete.rs
@@ -51,7 +51,7 @@ impl Shader for NeverComplete {
         None
     }
 
-    fn cell_filter(&self) -> Option<CellFilter> {
+    fn cell_filter(&self) -> Option<&CellFilter> {
         self.effect.cell_filter()
     }